Radical realignment of following UI

+++ dfinity_wallet/lib/ic_api/native_ic_api.dart
@@ -66,7 +66,7 @@ class PlatformICApi extends AbstractPlatformICApi {
   }
 
   @override
-  Future<void> increaseDissolveDelay({required BigInt neuronId, required BigInt additionalDissolveDelaySeconds}) {
+  Future<void> increaseDissolveDelay({required BigInt neuronId, required int additionalDissolveDelaySeconds}) {
     // TODO: implement increaseDissolveDelay
     throw UnimplementedError();
   }
